fix(ci): python-coverage-shards — split 1949-line outdated-resolution test over the 1500-line maintainability cap
test_pr_monitor_outdated_resolution.py grew to 1949 lines as the #547/#548
comment-keyed reconcile cases accreted, tripping
test_first_party_code_files_stay_under_line_limit (MAX_FIRST_PARTY_FILE_LINES
= 1500). Decompose it into a test_pr_monitor_outdated_resolution_parts package:
shared ReviewThread builders/stubs in _helpers.py, the PostgreSQL factory
fixture in conftest.py, and the 40 test cases split across two part files
(#473/#484 + helpers in part 1, #547/#548 reconcile in part 2). No test logic
changed; all 40 cases preserved.
